digitalOceanSpace_FileUpload/requirements.txt
# Following will be installed.

boto3
pymysql
python-dotenv
